Added local Subproject CMake module.
Subproject is a small CMake module or package that sets up a local system for managing CMake's External Projects. Adding a project by a name will search for a configuration file and if one is not found, a stubbed out one is generated. Until it is implemented it will generate developer warnings, but the configuration step of CMake will still complete.